BE/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
dnspython>=2.4.2
requests>=2.31.0